Switch from Helm-managed ingress to DuploCloud-managed ingress.
Benefits:
  Makes the ingress visible in the DuploCloud portal.
  Lets DuploCloud to manage the ingress DNS record.
  Lets DuploCloud manage network details like SG and subnets.

This includes some light refactoring and cleanup to keep its
changes from leaving behind vars that are now unused, etc.

This preserves the Helm values needed to use the chart to manage
the ingress in a comment with a note for users who decide to
use that approach (based on PR feedback).